Added a comment
I added a comment on the last line of the file which is just an example. But, example is an example and it should remain to be only taken an example. In reality, are examples not just examples ? - what if example becomes real, does it also mean its still an example ?